Partial testing for get_balances_in_date_range refactor
